feat(utils): add getCatalogItemDescriptor function to retrieve item path and descriptor based on item type
The getCatalogItemDescriptor function is added to the utils folder in the ui-dsc package. This function takes a CatalogItem object as input and uses pattern matching to determine the item type (plasmid or strain) and return the corresponding item path and descriptor. This function enhances the codebase by providing a clean and concise way to retrieve specific information based on the item type.