feat(ui-dsc): add getCatalogItemPathAndDescriptor utility function
The `getCatalogItemPathAndDescriptor` utility function is added to the `ui-dsc` package. This function takes a `CatalogItem` object as input and returns an object with `itemPath` and `itemDescriptor` properties based on the type of the `CatalogItem`. If the `CatalogItem` has a `name` property, it is considered a `plasmid` and the `itemPath` is set to "plasmids" and the `itemDescriptor` is set to the `name` property. If the `CatalogItem` has a `label` property, it is considered a `strain` and the `itemPath` is set to "strains" and the `itemDescriptor` is set to the `label` property. If the `CatalogItem` does not match any of these patterns, the `itemPath` and `itemDescriptor` are set to empty strings.
This utility function is added to provide a convenient way to extract the `itemPath` and `itemDescriptor` from a `CatalogItem` object in the UI of the `ui-dsc` package.