feat(HeaderRow.tsx): add support for plasmid items in the cart
The HeaderRow component now includes support for plasmid items in addition to strain items in the cart. The plasmidItemsAtom is imported and used to calculate the total number of items in the cart by combining the strainItems and plasmidItems arrays. The CartIcon component now receives the updated cartItems array as the items prop. This change allows for a more comprehensive representation of the items in the cart.