feat(ui-dsc): add OrderSummaryListItems component to display a list of items in the order summary
The new OrderSummaryListItems component is added to the project to display a list of items in the order summary. This component takes in a 'cart' object as a prop and renders a list of items with their descriptions, fees, and IDs. The component uses Material-UI components and fp-ts library functions for mapping over the items in the cart.