fix(ui-dsc): Update tests
Three test files have been updated to account for the addition of the
Plasmid catalog. The prop `strainItems` in CatalogRows has been updated
to `items`. The ContinueShoppingCard's has been changed to have two
buttons: one for the Strain Catalog and one for the Plasmid Catalog. The
tests for the ContinueShoppingCard have been updated to look for these
elements.