feat(AddPageEditor.tsx): add support for uploading images in the AddPageEditor component
The useUploadFileMutation hook is added to handle image uploads in the AddPageEditor component. The onImageUpload function is implemented to upload an image file using the uploadImage mutation and return the URL of the uploaded image. This enhancement allows users to upload images while creating or editing pages in the application.