feat(news-component, ui-common, ui-dsc): add mock for @logto/react in setup files
The changes introduce a new mock for the "@logto/react" library in the setup files of the news-component, ui-common, and ui-dsc packages. This mock provides a function that returns an object with a method to get an access token. This measure prevents several tests from failing that indirectly rely on @logto/react, which tries to import "react/jsx-runtime" which cannot be resolved in node.