feat(Dockerfile): add BUILD_STATE argument to Dockerfile for both frontpage and stock-center
The addition of the BUILD_STATE argument allows for more flexibility
during the build process, as it can be set to either "development" or
"production" depending on the desired environment. The .env file
renaming command now uses this BUILD_STATE argument, which makes the
Dockerfile more adaptable to different build environments.