fix(ci.yml): update workflow file used in docker build steps
The docker build steps in the GitHub Actions workflow were previously
using the 'docker-build-push-repo.yaml' file. This has been updated to
use the 'docker-build-push-mode.yaml' file instead. This change was
necessary to align with the updated workflow structure and to ensure the
correct execution of the docker build and push steps.