feat(.github/workflows): add staging.yml for staging integration
The new staging.yml file is added to the .github/workflows directory to
automate the build and push process for the Docker images of the
frontpage, stock-center, and publication services. This is done whenever
there is a push to the staging branch. This automation will help in
maintaining a consistent and error-free deployment process.