feat(FlexParagraphNode.ts): add FlexParagraphNode class to extend ParagraphNode for creating flexible paragraph nodes
The FlexParagraphNode class is added to extend the functionality of the ParagraphNode class. This new class allows for the creation of flexible paragraph nodes that can be styled with flexbox properties. The createDOM method is overridden to set specific styles for the flex container, making the paragraph node flexible and centered within its container.