feat(dicty-frontpage): add "@playwright/test" package to enable end-to-end testing with Playwright
The package.json file for the dicty-frontpage app has been updated to include the latest version of the "@types/node" package to ensure compatibility with the latest Node.js types. Additionally, the "@playwright/test" package has been added to enable end-to-end testing using Playwright, enhancing the testing capabilities of the application.