refactor(StrainDetailsCard.tsx): remove unnecessary in_stock variable and inStock prop
The in_stock variable was removed as it was not being used in the component. The inStock prop was also removed from the PhenotypeTable component as it was no longer needed after the removal of the in_stock variable. This cleanup improves code readability and removes unused variables and props.