feat(InsertLinkButton.tsx): add functionality to insert a link in the editor when the InsertLinkButton is clicked
The InsertLinkButton component now includes functionality to insert a link in the editor when the button is clicked. This is achieved by updating the editor state on click to toggle a link and insert the specified link text. This enhancement improves the user experience by providing a convenient way to add links within the editor.