ci(ci.yml): update Docker workflow actions to use main branch and simplify image specification
The GitHub Actions workflow now references the `docker-publish.yaml`
from the `main` branch instead of `docker-build-push-mode.yaml` from the
`develop` branch. This change ensures that the latest stable version of
the workflow is used. Additionally, the `image` key is used directly to
specify the Docker image name, simplifying the configuration by removing
the `dockerrepo` key.