chore(UnAuthorized.tsx): remove unused local image import and replace it with a remote image URL
The local image import "sadDicty" is removed as it is not being used in the component. Instead, a remote image URL is now used to display the image in the component. This change helps in reducing unnecessary imports and improves the loading time of the component by directly fetching the image from a URL.