test(SearchPhenotypeListItem.test.tsx): refactor test to improve readability and maintainability
The test for checking the presence of strain descriptor has been renamed to "expects strain descriptor" to better reflect its purpose. The test for displaying gene names has been refactored to use test.each for better readability and maintainability. Additionally, a new test has been added to ensure the publication link is rendered correctly. These changes aim to improve the clarity and organization of the test suite for the SearchPhenotypeListItem component.