feat(plopfile.js): add new templates and copies for Vite based dictybase applications
The changes were made to update the imports of fs and path modules to use the node: prefix for better clarity and to indicate that they are core Node.js modules. Additionally, new templates and copies were added to support Vite based dictybase applications, enhancing the project structure and functionality.