feat(utils): add ordByDate utility function for sorting dates in fp-ts library
The ordByDate utility function is added to the project to provide a way to compare dates using the fp-ts library. This function uses date-fns for date comparison and fp-ts for defining the ordering logic. It enhances the project by enabling sorting functionality based on dates in a functional programming paradigm.