feat(EditablePages): add EditableView component for displaying and editing content with an editable action bar
The EditableView component is added to the EditablePages feature to display content with an editable action bar. This component includes an ActionBar with a button to edit the content, showing the full name of the user who last updated the content and the time since the last update. The Editor component is used to display the content in a non-editable mode within a Container.